fix(guards): split the deployment-script guard and teach it the two shapes it went blind to (#14518)
The module reached 604 lines, four over the file-size ceiling. The reporting
layer (#14867) is an independent concern with its own detector, exemption table
and controls, so it moves to deployment_script_reporting_test; the block
extractor both layers share moves to deployment_script_scan. 604 -> 457, with
nothing dropped and no KNOWN_LARGE entry.

Splitting surfaced a real regression the discovery floors were already failing
on: validate_access_control.sh now builds each program in a quoted heredoc and
runs it through a run_python_check helper. Neither shape matched, so the guard
read exactly one block for that script -- the helper's own python3 -c "$1" --
and inspected none of the seven programs. The extractor learns both shapes and
drops shell indirections, which never name source it has not already read. Only
the two double-quoted shapes and the runner call feed the command-substitution
and fabrication detectors, because a quoted heredoc is not interpolated.
